Create challenges, data_Structures, and assets folders. Move array_reverse to challenges folder.  Add array_reverse solution image.
