+++ Challenges/array_reverse/array_reverse.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27703.2026
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "array_reverse", "array_reverse\array_reverse.csproj", "{7033BC35-D45C-47F8-A438-79E9026E3148}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{7033BC35-D45C-47F8-A438-79E9026E3148}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7033BC35-D45C-47F8-A438-79E9026E3148}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7033BC35-D45C-47F8-A438-79E9026E3148}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7033BC35-D45C-47F8-A438-79E9026E3148}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {F4957743-B743-4E6B-BA70-870B41DFBE3A}
+	EndGlobalSection
+EndGlobal
