--- array_reverse/README.md
@@ -1,6 +1,0 @@
-# Reverse an Array
-Given an array
-
-## Challenge
-
-## Solution
